the period finally is an intger YAAAAAAAY
